! Pooleli ! * Geneetiline algoritm olemas? * Muudetud mängusisu kasutamaks geneetilist algoritmi
